feat: move resolving detectors from names into `osvscanner`
